backend/app/requirements.txt
# requirements.txt
langgraph
langchain
langchain-anthropic
langchain-community
tavily-python
python-dotenv
langchain-groq